+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+# UNRELEASED
+
+- fix event id storage (@hertzcodes)